requirements.txt
# Deep Learning Frameworks
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.15.0

# Data Processing
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
pillow>=8.0.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Dimensionality Reduction
umap-learn>=0.5.0

# Utilities
tqdm>=4.62.0

# For Handling Files
kaggle>=1.5.12

# For System Operations
python-dotenv>=0.19.0

# Optional: For Jupyter Notebooks if needed
jupyter>=1.0.0
ipykernel>=6.0.0

peft>=0.5.0
accelerate>=0.20.0

torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
numpy>=1.19.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
scipy>=1.7.0
umap-learn>=0.5.0
tqdm>=4.60.0
Pillow>=8.2.0
Image